Rnadom Allied and Soviet Sides
We would like to see a "Random Allied/Soviet" Side in addition to the "Random" option under Side (Factions). This would allow players to get a random Soviet/Allied faction, as some times we'd like to play soviet & don't want to play Allied (& vice versa). This will also give a lot of variety to showmatches.