Update spawner.xdp
Previously the CowShot weapon(president ifv) was nerfed to deal 1 damage. Unfortunately there are dozens and dozens of survival maps which have this ifv mode for new units and this breaks all of them.

This change will restore the damage values but change the IFVMode to 2(GI gun inside IFV) for the president in order to fix a bunch of maps at once.