Correct the rubble palette for `caegyp01.shp`
In fact, this should be the graphic framework of `caegyp02.shp`, and as for the other three parts of the Great Pyramid, CAEGYP04-06 do not have SNOW versions, so CAEGYP01 does not have a SNOW version either. However, to avoid potential problems caused by this erroneous file, we still perform a simple processing on it.